Move files from views folder
